+++ Jenkinsfile.github
@@ -94,7 +94,7 @@ pipeline {
             environment {
                 PY_BIN_PATH = getPythonBinaryPath(PY_VERSION)
                 // If pylint rates the code lower than this, the build will fail
-                MIN_PYLINT_RATING = 9.2
+                MIN_PYLINT_RATING = 9.3
             }
             steps {
                 sh "make pylint"
