+++ src/csm_testing/tools/print_goss_json_results/duration_seconds.py
@@ -1,0 +1,53 @@
+#
+# MIT License
+#
+# (C) Copyright 2022-2024 Hewlett Packard Enterprise Development LP
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the "Software"),
+# to deal in the Software without restriction, including without limitation
+# the rights to use, copy, modify, merge, publish, distribute, sublicense,
+# and/or sell copies of the Software, and to permit persons to whom the
+# Software is furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR
+# OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+# ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+# OTHER DEALINGS IN THE SOFTWARE.
+#
+"""
+DurationSeconds class
+"""
+
+
+class DurationSeconds:
+    """
+    Ensures that when the entries are dumped as JSON in grok_exporter_logger,
+    the desired formatting of the duration in seconds is preserved.
+    """
+
+    def __init__(self, nanoseconds: int):
+        self.nanoseconds = nanoseconds
+        self.seconds = nanoseconds / 1000000000.0
+
+    def to_nanoseconds(self) -> int:
+        return self.nanoseconds
+
+    def __repr__(self) -> str:
+        # Print 9 decimal places because the minimum possible value is 0.000000001 (1 nanosecond)
+        return f"{self.seconds:.9f}"
+
+    def __str__(self) -> str:
+        return self.__repr__()
+
+    def to_json(self):
+        """
+        Used by the custom-JSON-encoding function in the grok_exporter_logger module
+        """
+        return self.__repr__()
